+++ packages/blocks/core/js/bootstrap.js
@@ -3,6 +3,7 @@ import {
 	bootstrapBlocksyBreadcrumbs,
 	bootstrapBlocksyAboutMe,
 	bootstrapBlocksyContactInfo,
+	bootstrapBlocksySocials,
 } from './third-party/bootstraps';
 import { bootstrapSocialLinksCoreBlock } from './wordpress/bootstraps';
 
@@ -12,4 +13,5 @@ export function blockeraBootstrapBlocks() {
 	bootstrapBlocksyBreadcrumbs();
 	bootstrapBlocksyAboutMe();
 	bootstrapBlocksyContactInfo();
+	bootstrapBlocksySocials();
 }