+++ packages/blocks/core/js/third-party/bootstraps.js
@@ -4,3 +4,4 @@ export * from './outermost-icon-block/bootstrap';
 export * from './blocksy-breadcrumbs/bootstrap';
 export * from './blocksy-about-me/bootstrap';
 export * from './blocksy-contact-info/bootstrap';
+export * from './blocksy-socials/bootstrap';