+++ packages/blocks/core/js/third-party/index.js
@@ -9,3 +9,4 @@ export * from './blocksy-tax-template';
 export * from './blocksy-widgets-wrapper';
 export * from './blocksy-about-me';
 export * from './blocksy-contact-info';
+export * from './blocksy-socials';